+++ tests/testthat/test-source_dir.R
@@ -1,13 +1,8 @@
 test_that("source_dir()", {
-  res <- source_dir("test_dir", pattern = "hello", chdir = TRUE, wrap = FALSE)
+  dir <- test_path("test_dir")
+  res <- source_dir(dir, pattern = "hello", chdir = TRUE, wrap = FALSE)
   expect_equal(res[[1]](), "Hello World")
 
-  res <- source_dir(normalizePath("test_dir"), pattern = "hello", chdir = TRUE, wrap = FALSE)
-  expect_equal(res[[1]](), "Hello World")
-
-  res <- source_dir("test_dir", pattern = "hello", chdir = FALSE, wrap = FALSE)
-  expect_equal(res[[1]](), "Hello World")
-
-  res <- source_dir(normalizePath("test_dir"), pattern = "hello", chdir = FALSE, wrap = FALSE)
+  res <- source_dir(dir, pattern = "hello", chdir = FALSE, wrap = FALSE)
   expect_equal(res[[1]](), "Hello World")
 })